Image Encryption and Decryption project in Java
Image Encryption/Decryption is a desktop application to encrypt or decrypt an image using a pass key. This application is built in Java using Swing and Awt. This project is build to encrypt or decrypt the image using a key. In this, I had to pass a key to encrypt the image and image will be encrypted. The same key to be passed is necessary to decrypt that image and the image will be decrypted at the same directory.